moved VM related methods to vm.go
